Switch to segmented database for sequence data
Reworks the sequence data storage to make use of the new
segmented_database framework to reduce the GPU memory footprint.
Eliminate explicit genome coordinate transforms by implementing an
abstraction for sequence data lookups.

Incidental changes include switching the chromosome index storage to
uint16, removing the sequence loading flags and eliminating one of the
alignment window structures in the main Firepony context (we now always
store alignment windows in chromosome coordinates).

+++ device/alignment_data_device.h
@@ -82,7 +82,7 @@ struct alignment_batch_device : public alignment_batch_storage<system>
         uint32 num_reads;
         uint32 max_read_size;
 
-        typename vector<system, uint32>::const_view chromosome;
+        typename vector<system, uint16>::const_view chromosome;
         typename vector<system, uint32>::const_view alignment_start;
         typename vector<system, uint32>::const_view alignment_stop;
         typename vector<system, uint32>::const_view mate_chromosome;
